requirements.txt
gspread==5.7.2
oauth2client==4.1.3
tk==0.1.0
python-dotenv==1.0.1
pyyaml==6.0.2
pillow==11.0.0